% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizr.R
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(x, minData = min(x), maxData = max(x), minNorm = 0,
  maxNorm = 1)
}
\arguments{
\item{x}{A vector of numbers}

\item{minData}{Data value to scale to minNorm}

\item{maxData}{Data value to scale to maxNorm}

\item{minNorm}{Value to scale minData to}

\item{maxNorm}{Value to scale maxData to}
}
\value{
A vector scaled with minData-maxData mapped to the range minNorm-maxNorm
}
\description{
Normalizes x from minData-maxData to the range minNorm-maxNorm
}
\examples{
normalize(0:10)  ## returns 0.0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 1.0
normalize(0:10, maxNorm = 100)  ## returns 0  10  20  30  40  50  60  70  80  90 100

}

